Update hub.py (#13)
Fixes a bug in function _to_room() (line 405)
The prior code assumed that the data for every channel includes the fields colorType, colorTitle and multiChannelComponent. On page 24 of ‘Accessing the Rako Hub from an external application v 0.3.0’, the API specification states in respect of colorType that 'This will only be set for the first channel in a multichannel sequence.' I found the field is not included in subsequent channels and the code failed.
Initially I changed the code to use the Python get function with a default value of None. After further thought, I changed it to store and reuse the value provided in the first channel for all subsequent channels. 
Initially I made this correction for just colorType but immediately got the same error for colorTitle. So I made the change for all three fields. The code then worked and created the light entities.